requirements-extra.txt
# testing/linting-related dependencies
black==22.12.0
coverage==7.1.0
flake8==6.0.0
gitlint==0.18.0
isort==5.12.0
mock==5.0.1
mypy==0.991
pre-commit==3.0.2
pytest==7.2.1
pytest-cov==4.0.0
pytest-django==4.5.2
coverage==7.1.0
